% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set}
\format{
A data frame with 144 rows and 6 variables:
\describe{
  \item{RIVER}{river}
  \item{MILE}{mile}
  \item{SPECIES}{species}
  \item{LENGTH}{length}
  \item{WEIGHT}{weight}
  \item{DDT}{ddt}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
ddt
}
\description{
A data set from intro2R
}
\keyword{datasets}
